Refactor memory dump viewer into modular architecture with sidebar navigation
