+++ src/S7Tools/Views/Pages/FileMemoryDumpView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace S7Tools.Views.Pages;
+
+public partial class FileMemoryDumpView : UserControl
+{
+    public FileMemoryDumpView()
+    {
+        InitializeComponent();
+    }
+}
