+++ src/S7Tools/Views/Pages/MemoryDumpSidebarView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace S7Tools.Views.Pages;
+
+public partial class MemoryDumpSidebarView : UserControl
+{
+    public MemoryDumpSidebarView()
+    {
+        InitializeComponent();
+    }
+}
