+++ src/S7Tools/Views/Pages/StreamedMemoryDumpView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace S7Tools.Views.Pages;
+
+public partial class StreamedMemoryDumpView : UserControl
+{
+    public StreamedMemoryDumpView()
+    {
+        InitializeComponent();
+    }
+}
